Add files via upload
Father forgive me for I have sinned. 
What it does: 
-gets two positions and is able to hold onto them with the right click
-does not overwrite it unless you left-click and do it again. 
-gets the positions to the functions that move in mainwindow
What it does not do:
-Move what is selected. 

it has every step done but it just does not move the things that  you want or anything I can't find out what in paticular it is missing I tryed making my own but just wound up breaking all of the other shifting functions I left in things that will not work in as comments and deleted stuff that probably was not useful.