Add files via upload
I keep getting errors trying to resolve the list of qcolors I had it exactly how the rest of the functions worked previously and it still did not work out maybe the problem is something I just can't see right now but I need sleep I can't keep working. Also making a new window dialog would be nice but I don't really remember how we wanted the user to make a stamp. I tried to make it so that they would make a selection the click the savesprite and it would do the action, thus bringing up the save file dialog but that part gave me errors and I can't find why.